ipmiPowerIndicator Module
